tehtävä 4.21

+++ osa4/controllers/blogs.js
@@ -8,18 +8,10 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-/*const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}*/
 
 blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
 
-    //const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid' })
@@ -46,9 +38,23 @@ blogsRouter.post('/', async (request, response, next) => {
     }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+    try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return response.status(401).json({ error: 'token invalid' })
+        }
+
+        const blogToBeDeleted = await Blog.findById(request.params.id)
+        if (blogToBeDeleted.user.toString() !== decodedToken.id) {
+            return response.status(401).json({ error: 'user not authorized to delete this blog' })
+        }
+        else {
+            await Blog.findByIdAndDelete(request.params.id)
+            response.status(204).end()
+        }  }
+    catch (error) {
+        next(error)}
 })
 
 blogsRouter.put('/:id', async (request, response) => {
